requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
psycopg[binary]==3.3.1
alembic==1.14.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2  # pin for passlib compatibility (__about__ removal in bcrypt 4.x)
python-multipart==0.0.21
pytest==9.0.2
